Fix full flavor remove mdx and mdd

+++ lib/repositories/dictionary.dart
@@ -13,6 +13,7 @@ import "package:dict_reader/dict_reader.dart";
 import "package:drift/drift.dart";
 import "package:file_selector/file_selector.dart";
 import "package:flutter/material.dart";
+import "package:flutter/services.dart";
 import "package:go_router/go_router.dart";
 import "package:html_unescape/html_unescape_small.dart";
 import "package:mime/mime.dart";
@@ -244,7 +245,7 @@ class Mdict {
 
     await dictionaryListDao.remove(path);
 
-    if (Platform.isAndroid) {
+    if (Platform.isAndroid && appFlavor != "full") {
       final mdxFile = File("$path.mdx");
       await mdxFile.delete();
